mock require context shenanigans
