+++ shared-data/js/__tests__/fixtures/labware/irregularLabwareExample1.json
