+++ shared-data/js/__tests__/fixtures/labware/labwareExample2.json
