+++ shared-data/js/labwareTools/__tests__/createLabware.test.js
@@ -3,8 +3,8 @@ import range from 'lodash/range'
 
 import { createRegularLabware } from '../index.js'
 
-import exampleLabware1 from '../../__tests__/fixtures/labwareExample.json'
-import exampleLabware2 from '../../__tests__/fixtures/labwareExample2.json'
+import exampleLabware1 from '../../__tests__/fixtures/labware/labwareExample.json'
+import exampleLabware2 from '../../__tests__/fixtures/labware/labwareExample2.json'
 
 jest.mock('../assignId', () => jest.fn(() => 'mock-id'))
